Rename zika func module more appropriately. Small fixes in the methods.
